fixed opencl int mismatch with size_t

+++ src/layers_cl/eltwise_layer_cl.h
@@ -40,10 +40,10 @@ class EltwiseLayerCL: public Layer<Dtype>
 
 
     private:
-        size_t output_height;
-        size_t output_width;
-        size_t output_channels;
-        size_t channel_block_size;
+        int output_height;
+        int output_width;
+        int output_channels;
+        int channel_block_size;
         bool fuse_relu;
 
 };
